Added return value for delete profile

+++ ovs/services/profile_service.py
@@ -45,7 +45,10 @@ class ProfileService:
         Deletes a user's profile information
         """
         resident = ResidentService.get_resident_by_id(resident_id).one_or_none()
+        if resident is None:
+            return False
         profile = resident.profile
         picture_id = profile.picture_id
         ProfilePictureService.delete_profile_picture(picture_id)
         db.delete(profile)
+        return True
